Move array hasher into separate file
